Check if server is running before starting a new server

+++ start.sh
@@ -15,57 +15,74 @@ rm jzedit.desktop-bak
 # Make it runable again
 chmod +x jzedit.desktop
 
-echo Starting the server ...
-cd server
-serverArg="--loglevel=6 --username=admin --password=admin"
 
-# Check if sudo/root
-if [ "$EUID" -ne 0 ]
-  # Not root, use a different port then 80
-  then serverArg="$serverArg --port=8080"
-fi
-
-# Check if we should use nodejs or just node
-if type nodejs >/dev/null
-then NODE=nodejs
-
-elif type node >/dev/null
-then NODE=node
-
-else 
-	echo "We need Node.JS! Download it from www.nodejs.org, or # sudo apt-get install nodejs"
-	exit
-fi
-
-CMD="$NODE server.js $serverArg"
-
-# Start server in background
-# Ignore stdout and send stderr to server.err
-$CMD &
-
-PID=$!
-# The pid is only correct if the server started successfully!
-# Don't use the pid to check if the server is running or not
-
-# Give the server a chance to exit if there's something wrong
-sleep 1
-
-# Make sure the server is running
-if ! (ps ax | grep "[n]${CMD:1}" > /dev/null) then
-    echo "Could not start the server!" >&2
-else
-	echo Server started! Starting the client ...
-
-	cd ..
-
-
-	# Start the client (editor
+function startClient {
+	echo Starting the client ...
+	# Start the client (editor)
 	#runtime/nwjs-v0.12.3-linux-x64/nw . --remote-debugging-port=57341
 	runtime/nwjs-v0.12.3-linux-x64/nw . --disable-lcd-text > /dev/null 2>&1
+}
 
 
+# Check if the server is already running
+if (ps aux | grep nodej*s*.*server\.js > /dev/null) then
+	echo Looks like the server is already running!
 
-	# Kill the server when client exit
-	kill $!
+	startClient
 
+else
+
+
+	echo Starting the server ...
+	cd server
+	serverArg="--loglevel=6 --username=admin --password=admin"
+
+	# Check if sudo/root
+	if [ "$EUID" -ne 0 ]
+	  # Not root, use a different port then 80
+	  then serverArg="$serverArg --port=8080"
+	fi
+
+	# Check if we should use nodejs or just node
+	if type nodejs >/dev/null
+	then NODE=nodejs
+
+	elif type node >/dev/null
+	then NODE=node
+
+	else 
+		echo "We need Node.JS! Download it from www.nodejs.org, or # sudo apt-get install nodejs"
+		exit
+	fi
+
+	CMD="$NODE server.js $serverArg"
+
+	# Start server in background
+	# Ignore stdout and send stderr to server.err
+	$CMD &
+
+	PID=$!
+	# The pid is only correct if the server started successfully!
+	# Don't use the pid to check if the server is running or not
+
+	# Give the server a chance to exit if there's something wrong
+	sleep 1
+
+	# Make sure the server is running
+	if ! (ps ax | grep "[n]${CMD:1}" > /dev/null) then
+	    echo "Could not start the server!" >&2
+	else
+		echo Server started!
+
+		cd ..
+
+
+		startClient
+
+
+		# Kill the server when client exit
+		kill $!
+
+	fi
 fi
+
